doc update

+++ src/tir/schedule/transform.cc
@@ -221,6 +221,7 @@ void LeafBlockRemovalPlan(const ScheduleState& self, const StmtSRef& leaf_block_
   }
   if (const auto* block = sref->StmtAs<BlockNode>()) {
     auto body = block->body;
+    // Peel off AllocateConst nodes at the beginning of the block body.
     std::vector<const AllocateConstNode*> allocs;
     while (const auto* alloc = body.as<AllocateConstNode>()) {
       allocs.push_back(alloc);
@@ -229,8 +230,9 @@ void LeafBlockRemovalPlan(const ScheduleState& self, const StmtSRef& leaf_block_
     if (const auto* seq = body.as<SeqStmtNode>()) {
       ObjectPtr<BlockNode> n = make_object<BlockNode>(*block);
       auto new_seq = RemoveFromSeqStmt(GetRef<SeqStmt>(seq), GetRef<Stmt>(last_stmt));
+      // Re-attach AllocateConst nodes
       auto new_body = new_seq;
-      for (int i = 0; i < allocs.size(); ++i) {
+      for (int i = 0; i < static_cast<int>(allocs.size()); ++i) {
         auto alloc = allocs[allocs.size() - 1 - i];
         new_body = AllocateConst(alloc->buffer_var, alloc->dtype, alloc->extents, alloc->data,
                                  new_body, alloc->annotations, alloc->span);
